Add option to disable portal support and switch to X11
If the user does not have a mutter with remote desktop enabled, then it
may be viable to use an X11 image source instead. This is a compile time
option only though, as we generally want to rely on the remote desktop
protocol and possibly portal in the future.

Adding this makes it possible to try on machines that don't have support
in mutter (e.g. ubuntu).