wfd: Fix some parameters
This changes a few parameters. The only change in here that actually
fixes things in this case is forcing the slice count to be 1. There is
actually another underlying bug (will be fixed in following commits)
that caused a too high slice count to be set.

However, it doesn't hurt much to be on the safe side.