wfd: Generate QOS events for realtime and adjust options
Sometimes the encoder may not run in real-time. In particular the
openh264 encoder element will start having a lot of buffer in the queue
rather than dropping input buffers.

Generate QOS events if buffers start coming in late, to force the codec
to drop frames and ensure playback is smooth (without audio dropping
out).

Also adjust some default options, including limitting the video bitrate
so that the wifi link is not over-saturated.