wfd-p2p: Set IP configuration for new connection
Usually the defaults are good enough for us. The exception is that we
may get routes through DHCP if the remote peer becomes the group owner.
We really don't want to set default routes for the P2P connection.

Set the appropriate flags in the IPvX configuration to avoid this from
happening. Also explicitly allow IPv6 configuration to fail.

Fixes: #16