portal: Fall back to X11 for all errors and improve error reporting
It seems that there are users out there who have the XDG portals installed,
but no support for screencasting in mutter. Apparently, we receive
errors from the portal in that case, which resulted in us giving up
right away rather than trying a fallback.

Obviously these users should fix their setup and compile mutter with
pipewire support (so that screencasting is available). But we can try to
handle it better and simply fall back to X11 rather than erroring out
without any further feedback.

Also improve the error reporting a bit, so that we can see where exactly
the error came from in the future.

Fixes: #17