wfd: Always touch session for incoming packets
It turns out that some adapters (specifically MS v2 adapters) do not
reply with the session ID for the keep-alive packet. In this case the
gstreamer RTSP server will not automatically touch the GstRTSPSession
and we will run into the session timeout.

Fix this by simply touching all session for the client, for any received
response.

See: #20