screencast-portal: Fix error reporting in portal creation
With ce8d6e1f (screencast-portal: Fix cancellation handling) the
internal cancellation of tasks would cause G_IO_ERROR_CANCELLED to be
reported externally. This must not happen though, because the
surrounding code will ignore the result as it thinks that the
cancellation happened due to cleaning up for itself.

Fixes: #76