wfd: Add (nasty) workaround for clients that do not send Session header
Some clients do not send a session header. Unfortunately, there is no
proper way to hook into the gstreamer side to work around this.

There is a way though, which is by abusing an authenticator. But for
that to work we need to register the authenticator after the client has
connected (otherwise we would ask for credentials).

Fixes: #78